Заказы в гильдии -Добавлена возможность смотреть заказы в гильдии -Добавлена возможность добавлять заказы админу -Добавлены соответсвующие репозитории,Dto,сервисы

+++ src/main/java/com/example/kptc_smp/repositories/RoleRepository.java
@@ -1,13 +1,12 @@
 package com.example.kptc_smp.repositories;
 
 import com.example.kptc_smp.entitys.Role;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByName(String name);
 }
